chatbot_service/requirements.txt
fastapi
openai==0.28
uvicorn
python-dotenv
sentence-transformers
numpy
requests
